PERF: Reject anonymous requests with 503 upon aggressive queuing
This commit adds a default protection to all Discourse app which will
turn away all anonymous requests that has queued for more than the duration configured
by the `reject_anonymous_min_queue_seconds` global setting, 1 second
by default. A queue time of 1 second is a sign that the app is very
overloaded so we are opting to turn away anonymous requests when that
happens.

While there may be better solutions like introducing an adaptive
concurrency limit somewhere in the stack, we like this solution for its
simplicity. We may consider the more complex solutions in the future but
the time is not now.

Other significant changes in this commit:

1. Moved setting of the `REQUEST_QUEUE_SECONDS` key in `env` to
   `Middleware::ProcessingRequest`.
2. Introduction of the `reject_anonymous_min_queue_seconds` global
   Setting.