FIX: Allow moderators to unhide posts using acting_user context (#36512)
We were not acting as moderators when unhide, making it impossible for
moderators
to unhide posts that violated a post's validation rules.

(eg: too many links, not allowed to have embedding, etc)

Note: this still leaves the post "uneditable" by the actual user, but at
least
give the moderator a way of allowing something banned by the rules.


https://meta.discourse.org/t/embedded-media-post-allowed-groups-conflicts-with-post-moderation-workflow-unhide-fails-with-activerecord-recordinvalid/390575

---------

Co-authored-by: Krzysztof Kotlarek <kotlarek.krzysztof@gmail.com>