FIX: ReFix the loading loop when a topic is removed. (#36410)
In #35409, a fix has been implemented to deal with the loading loop when
a topic is removed.

However, that PR hasn't fully solved that problem. In the Glimmer 
Post Stream, the fetch API is `findPostsByIds`->`loadIntoIdentityMap` 
rather than the original `fetchNextWindow`, causing a regression.

This PR adds the same logic to the `loadIntoIdentityMap` function, so
that the logic works as well in the post-stream.

Co-authored-by: Sérgio Saquetim <1108771+megothss@users.noreply.github.com>