DEV: Remove deprecated trust level access settings (#36414)
### What is this change?

Some time back we went through an exercise to turn trust-level based
access to group-level based for better configurability. This involved
deprecating, mapping, and replacing site settings. We marked the old
settings for removal in 3.3, which was released over a year ago now.

This PR removes the deprecated settings, their descriptions, the mapping
helpers, and updates a few test cases we missed in the original
conversion.

### Due diligence:

- [x] Search through logs for recent deprecation warnings. (@Drenmi)
- There are some false positives due to `UserMerger` looping through all
site settings (including deprecated ones.)
- [ ] Search through public and private repos for latent references.
(@Drenmi)

---------

Co-authored-by: Martin Brennan <martin@discourse.org>