FEATURE: Reset the 'show original content' status when the user changes content language (#36620)
When a user changes the language via the switcher, we want to reset the
"Show Original Content" state because, well, the user obviously wants to
see content in a different language.

This came about because the button is not very obvious in indicating its
status as well, resetting the status will help.


t/170623/12