DEV: Allow multiple upload fields in object site setting (#36605)
When multiple upload types were included in a object site setting, they
were not given unique identifiers. This is problematic as all inputs
were treated as one, meaning you could upload a asset to UPLOAD_FIELD_1,
however uploading a asset to UPLOAD_FIELD_2 would override the first
field.

This PR concats the name of the setting to the id of the upload field to
create a unique identifier.