PERF: Improve performance of DiscourseTagging.hidden_tag_names (#36636)
Improve performance of `DiscourseTagging.hidden_tag_names` by using its
own query.

Previously it added `where.not(id: ` to the query from
`DiscourseTagging.visible_tags`. This simplifies the hidden_tag_names
query by just directly running the inverse query, to find names of tags
which have at least one TagGroupPermission but do not have any for
permitted tag groups.

This query gets run every time tags are displayed on a page, and the
goal here is to improve performance in the case tracked in internal
topic /t/168944. On that customer's site, this reduces the query time
from ~11ms to ~5ms.