FIX: Related setting toggle on what's new page for admins (#36646)
Followup d4ac43e605d4674eece7056df18a20a4a8308cb0

A regression was introduced in the above commit where an `enabled`
param was added to `UpcomingChanges::Toggle`, which is also used
by the what's new feature feed items. When using the toggle, admins
would get this message:

> An error occurred: Enabled is not included in the list

This regression was not caught because there was no system spec
coverage for this toggle in the what's new feed. This commit
fixes the issue and adds coverage.