DEV: Extract unicorn server PID when ArgumentError raised (#36649)
When the rails server is already running, attempting to start
it again was giving an ArgumentError and outputting the backtrace.
When you are working locally, all you want here is the PID so you
can kill it.

I altered the error message to this in this case:

```
Unicorn is already running!

kill -9 150044 to stop the existing unicorn master process, then try again.
```

So you can easily copy and paste the command to stop the existing
server.

The old error:

<img width="1711" height="243" alt="image"
src="https://github.com/user-attachments/assets/18d5b25b-5da9-45b0-941e-6ee353423cf3"
/>