UX: Prevent `HorizontalOverflowNav` component from scrolling viewport
Whenever the `HorizontalOverflowNav` component is rendered, we call the
`scrollIntoView` on the current active element to ensure that the link
which is currently active will be shown to the user. However, we were
calling `scrollIntoView` with the `container` option defaulting to `all`
which impacts all scrollable containers.

Per mozilla docs,

> All scrollable containers are impacted (including the viewport).

This is not ideal as it can cause unintended scrolls on the viewport
when there are other scrollable containers on the user's screen thus
resulting in a bad UX. To resolve this, this commit adds the `container:
"nearest"` option when we call `scrollIntoView` which ensures that only
the nearest scrollable container to the active element is impacted.

This problem was discovered while I was looking at review queue index
page and noticed that my scroll position was always lost when new
reviewable items are added to the page. It turns out that the review
queue index page renders the `reviewable-refresh/item.gjs` component
which then renders the `horizontal-overflow-nav.gjs` component.