FIX: Address race condition in `PostStream` while loading posts (#36744)
Convert `findPostsByIds` and `loadIntoIdentityMap` to async functions in
`PostStream` to resolve a race condition when loading posts. This
ensures proper sequencing of asynchronous operations and improves
reliability in loading post data.

This PR addresses the issue reported in
https://meta.discourse.org/t/the-web-page-continuously-loads-previous-posts-and-automatically-scrolls-up-quickly/391192

There are no tests added because I can't find a way to reliably
reproduce the issue in tests and verify it's properly fixed.