FIX: Edit sidebar tags navigation model "selected" filter not working
What is the problem?

When a user does not have any tags configured to be displayed in the
sidebar, selecting the "selected" filter in the edit sidebar tags
navigation model was not displaying an empty state. Instead, it
displayed all the tags even though those tags were unselected.

The above behaviour is due to the fact that the client side was not
properly setting the `only_tags` param to the `/tags/list.json` endpoint
when the user has not selected any tags in the modal. As a result, the
endpoint ended up returning all the tags that are visible to the user.

What is the fix?

On the client side, we just have to avoid fetching tags when we already
know that the user has not selected any tags.