FEATURE: Introduce tag localizations with API, without UI yet (#36754)
This commit introduces tag localizations... but only on the API. We can go
down the stack:

### tag localizations controller
This adds the ability to GET, PUT, DELETE a tag localization, and checks
if the user can see the tag (tag group permissions) and localize based
on the `content_localization_allowed_groups` site setting. This also
adds each accompanying
- serializer
- routes
- guardian
- service

Note, this controller was created but I am not yet sure if we will be
using it in tags settings. Category settings make use of a big endpoint
that involves ALL category settings, so updating category localizations
is part of updating a category's settings. We will know more once I
implement the tags settings page that will come after slugs are
introduced.

### ai
On the AI side this adds the required jobs (backfill detection, backfill
localization), and associated services (tag candidates, localizer).

Localization for tags involve both `name` and `description`, and uses
the existing `ShortTextTranslator` and its quota.

_____

Reviewer note: This will look very similar to the other models, but I
chose this duplication over small if-elses scattered everywhere. We can
def revisit dedup if needed.