FIX: ensures moderators can see category permissions (#36772)
We could potentially return limited information to moderators who
shouldn't have access to it.

We considered this as FIX and not SECURITY as this endpoint is limited
to staff only and is only returning group names. Also you would need to
know the category ID as it wouldn't show in the UI if you can't access
it.