DEV: Make changes for docker_manager Pitchfork compatibility (#36456)
This brings the necessary changes to allow the `docker_manager` plugin
to work properly with Pitchfork (see
https://github.com/discourse/docker_manager/pull/293).

Three things are needed:
- Tell our MessageBus config to not return a 429 if a request takes more
than 100 ms while the Pitchfork server is restarting.
- Update the Pitchfork config to allow it to reuse its own port (this
allows us to start a new Pitchfork server while the old one is still
running)
- Update the `unicorn_launcher` script to restart the Pitchfork server
when it receives a USR2/HUP signal. This way, it doesn’t complicate
things too much.