A11Y: remove role="button" from solved quote title (#36783)
`role="button"` isn't needed because it's a `div` without `tabindex`,
which isn't reachable via keyboard navigation. This means the role
creates a weird situation where screenreaders may announce "button" but
can't actually focus it for use.

This quote expansion behavior that is added to the title *is* a nice
affordance to create a larger click area for mouse users... technically
we shouldn't nest links within this clickable title area like this, but
it works ok as-is. When testing with a screenreader we just navigate to
the links (username and post) as well as the "expand" and "jump to post"
button present on the right... so all the functionality still exists
anyway.

<img width="1248" height="154" alt="image"
src="https://github.com/user-attachments/assets/74853e9e-febe-4692-b3b3-ea77a1bc1b6f"
/>