FIX: Delete other posts in a topic shouldn't unanswer the answer post (#36800)
There exists a potential bug in discourse-solved plugin, when any of the
post in the answered topic is removed(trigger `post-destroyed`),
**whether the removed post is marked as the answer or not**, the marked
answer will be unanswered.

It seems weird because once someone removes an unrelated post in a
topic, the topic owner will have to reselect an answer, or they will be
notified by the discobot.

This commit fix this by adding a discrimination logic: only when the
post marked as answer is removed will `unanswer` be triggered, and the
removal of other posts won't do that.

Before:


https://github.com/user-attachments/assets/f67fff80-9c63-4762-a47b-4bf30894d4a4

After:


https://github.com/user-attachments/assets/1729c9f3-948d-45ed-a00e-8931b4ad2089