FIX: Disable load_more_replies button until more replies are loaded (#36816)
In PR #35136, a debounce logic is added to the reply button. However, a
similar issue occurs in the "load_more_replies" button as well.

This commit fix the issue by adding disable logic to the
`load_more_replies` button, so that the button is also disabled when
loading, and this issue is solved.

Before:


https://github.com/user-attachments/assets/7700a70d-3d8d-4837-8286-1f3bb8fc3006

After:


https://github.com/user-attachments/assets/f12bc001-d9c9-46d6-93a0-2f57d8e1584a