DEV: Resupport running `multisite:migrate` concurrently
Previously, we were using multiple threads to run migrations for
multisite in parallel. However, it was discovered that `db:migrate` is
not threadsafe and was not designed to be. As a result,
`multisite:migrate` was hardcoded to only support spinning up a single
thread.

This new attempt Uses the `parallel` gem with `in_processes` to run
migrations concurrently using multiple processes. Concurrency defaults to 2 but can be configured via the `DISCOURSE_MULTISITE_MIGRATE_CONCURRENCY` env. Example:

`DISCOURSE_MULTISITE_MIGRATE_CONCURRENCY=5 bin/rails multisite:migrate`