FIX: Update topic excerpts when the first post is localized (#36885)
When localizing a pinned topic, the excerpt shown in the topic list
wasn't updating - only the title would change when switching languages.
This was because the `TopicLocalization.excerpt` field was not being
updated.

The excerpt needs to be set in two scenarios depending on which gets
created first:
- When a post localization is created/updated for the first post, update
the topic localization's excerpt
- When a topic localization is created and a post localization already
exists, populate the excerpt from it

Meta:
https://meta.discourse.org/t/localizing-topic-does-not-seem-to-localize-excerpt-pinned-on-topic-list/392232