UX: Modifications for starred channels behavior
# :mag: Overview
In this update we update the behavior for starred channels. In particular for mobile/drawer modes. With this change, channel row items should always be shown in their original tab despite being starred. Also the indicators should be shown on the channel row item in both places. However, the indicator on the footer tabs should only be shown on the starred tab if a starred channel has unreads.