FIX: Load categories in edit navigation menu modal in hierarchical order
What is the problem?

Before this commit, pagination of categories in the edit navigation menu
modal only happened for the root categories (category without a parent).
child/child-child categories were sideloaded, with a limit of 5, alongside its parent's
category. When a parent category has more than 5 child categories, the
UI displays a "show more" button that allows a user to load more child
categories for a parent. However, this approach has been fraught with
complexity and many bugs that we've decided to adopt a different
approach.

What is being changed?

Instead of sideloading a limited set of child categories alongside
parent categories, we can instead paginate through the entire category
tree by flattening the tree by organizing the categories in a
hierarchical order where child categories are always positioned directly
after its parent category in the list. Not only is the SQL query for the
generating this list easier to implement, it is also easier for us to
paginate through the list leading to significantly less complexity on
the server and client side.