FIX: Respect title_count_mode preference in chat (#36930)
Follow-up to https://github.com/discourse/discourse/pull/36652

The linked PR adds logic to show a counter in the document title when
chat messages arrive to a channel that's left open in a background tab.
This behavior is consistent with the topic page where the document title
also shows a counter when posts are made. We do have an existing user
preference (Preferences > Interface > Background page title displays
count of:) to disable the document title counter for topics, however,
chat currently shows counter in the document title regardless of that
user preference. This PR makes chat respect the user preference.